backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
google-generativeai
python-multipart
Pillow
youtube-transcript-api